Delt inn i to service klasser
