+++ apps/tp-service/src/main/java/no/nav/testnav/apps/tpservice/provider/rs/SyntetiseringsController.java
@@ -2,9 +2,10 @@ package no.nav.testnav.apps.tpservice.provider.rs;
 
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import no.nav.testnav.apps.tpservice.database.models.TForhold;
 import no.nav.testnav.apps.tpservice.database.multitenancy.TenantContext;
 import no.nav.testnav.apps.tpservice.provider.rs.request.SyntetiseringsRequest;
-import no.nav.testnav.apps.tpservice.service.TpService;
+import no.nav.testnav.apps.tpservice.service.SyntService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,13 +15,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import java.util.List;
 
 @RestController
 @RequestMapping("api/v1/syntetisering")
 @RequiredArgsConstructor
 public class SyntetiseringsController {
 
-    private final TpService tpService;
+    private final SyntService syntService;
 
     @ApiOperation(value = "Dette endepunktet kan benyttes for å generere syntetiserte ytelser på tilfeldige personer i en gitt avspillergruppe som er definert i TPS-Forvalteren.")
     @PostMapping(value = "/generer")
@@ -28,16 +30,16 @@ public class SyntetiseringsController {
             @RequestBody @Valid SyntetiseringsRequest request
     ) {
         TenantContext.setTenant(request.getMiljoe());
-        tpService.syntetiser(request);
+        syntService.syntetiser(request);
         return ResponseEntity.ok().build();
     }
 
     @ApiOperation(value = "Dette endepunktet kan benyttes for å hente ut alle forhold i et gitt miljø.")
     @GetMapping(value = "/forhold/{miljoe}")
-    public ResponseEntity<Object> getForhold(
+    public List<TForhold> getForhold(
             @PathVariable String miljoe
     ) {
         TenantContext.setTenant(miljoe);
-        return ResponseEntity.ok(tpService.getForhold());
+        return syntService.getForhold();
     }
 }
